_pkg.sh: honor `CW_PKG_NODELETE` with `CW_NOPKG` [ci skip]

+++ _pkg.sh
@@ -105,7 +105,9 @@ create_pkg() {
 }
 
 if [ "${CW_NOPKG:-}" = '1' ]; then
-  rm -r -f "${_DST:?}"
+  if [ "${CW_PKG_NODELETE:-}" = '1' ]; then
+    rm -r -f "${_DST:?}"
+  fi
   exit
 fi
